requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

certifi==2020.6.20
waitress==2.0.0
chardet==3.0.4
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.5
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.0
requests==2.24.0
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
torch==1.10.1+cpu
torchvision==0.11.2+cpu
typing-extensions==4.0.1
Werkzeug==2.0.2
